NFTablesEx.RuleBuilder (NFTablesEx v0.4.0)

Fluent API for building nftables rules.

This module provides an intuitive, chainable interface for building firewall rules using nft syntax strings internally.

Quick Example

alias NFTablesEx.RuleBuilder

# Drop packets from specific IP
RuleBuilder.new(pid, "filter", "INPUT")
|> RuleBuilder.match_source_ip("192.168.1.100")
|> RuleBuilder.log("BLOCKED IP: ")
|> RuleBuilder.drop()
|> RuleBuilder.commit()

Common Patterns

# Accept established/related connections
RuleBuilder.new(pid, "filter", "INPUT")
|> RuleBuilder.match_ct_state([:established, :related])
|> RuleBuilder.accept()
|> RuleBuilder.commit()

# Rate limit SSH connections
RuleBuilder.new(pid, "filter", "INPUT")
|> RuleBuilder.match_dest_port(22)
|> RuleBuilder.rate_limit(10, :minute)
|> RuleBuilder.accept()
|> RuleBuilder.commit()

See Also

NFTablesEx.Policy - Pre-built common policies

NFTablesEx.Rule - Low-level rule operations

Types

t()

@type t() :: %NFTablesEx.RuleBuilder{
  chain: String.t(),
  expr_list: [map()],
  family: atom(),
  pid: pid(),
  table: String.t()
}

Functions

accept(builder)

See NFTablesEx.RuleBuilder.Verdicts.accept/1.

commit(builder)

@spec commit(t()) :: :ok | {:error, term()}

Commit the rule to the kernel.

Sends the structured JSON expression list directly to the kernel via the libnftables JSON API.

Returns :ok on success, {:error, reason} on failure.

Examples

# Build and commit a rule
RuleBuilder.new(pid, "filter", "INPUT")
|> RuleBuilder.match_source_ip("192.168.1.100")
|> RuleBuilder.drop()
|> RuleBuilder.commit()

continue(builder)

See NFTablesEx.RuleBuilder.Verdicts.continue/1.

counter(builder)

See NFTablesEx.RuleBuilder.Actions.counter/1.

decrement_hoplimit(builder)

See NFTablesEx.RuleBuilder.Actions.decrement_hoplimit/1.

decrement_ttl(builder)

See NFTablesEx.RuleBuilder.Actions.decrement_ttl/1.

dnat_to(builder, ip)

See NFTablesEx.RuleBuilder.NAT.dnat_to/2.

dnat_to(builder, ip, opts)

See NFTablesEx.RuleBuilder.NAT.dnat_to/3.

drop(builder)

See NFTablesEx.RuleBuilder.Verdicts.drop/1.

duplicate_to(builder, interface)

See NFTablesEx.RuleBuilder.Verdicts.duplicate_to/2.

flow_offload(builder)

See NFTablesEx.RuleBuilder.Verdicts.flow_offload/1.

flow_offload(builder, opts)

See NFTablesEx.RuleBuilder.Verdicts.flow_offload/2.

goto(builder, chain_name)

See NFTablesEx.RuleBuilder.Verdicts.goto/2.

increment_hoplimit(builder)

See NFTablesEx.RuleBuilder.Actions.increment_hoplimit/1.

increment_ttl(builder)

See NFTablesEx.RuleBuilder.Actions.increment_ttl/1.

jump(builder, chain_name)

See NFTablesEx.RuleBuilder.Verdicts.jump/2.

limit_connections(builder, count)

See NFTablesEx.RuleBuilder.CTMatching.limit_connections/2.

log(builder, prefix)

See NFTablesEx.RuleBuilder.Actions.log/2.

log(builder, prefix, opts)

See NFTablesEx.RuleBuilder.Actions.log/3.

masquerade(builder)

See NFTablesEx.RuleBuilder.NAT.masquerade/1.

masquerade(builder, opts)

See NFTablesEx.RuleBuilder.NAT.masquerade/2.

match_ah_spi(builder, spi)

See NFTablesEx.RuleBuilder.AdvancedMatching.match_ah_spi/2.

match_arp_operation(builder, operation)

See NFTablesEx.RuleBuilder.AdvancedMatching.match_arp_operation/2.

match_cgroup(builder, cgroup_id)

See NFTablesEx.RuleBuilder.AdvancedMatching.match_cgroup/2.

match_connmark(builder, mark)

See NFTablesEx.RuleBuilder.CTMatching.match_connmark/2.

match_ct_bytes(builder, op, bytes)

See NFTablesEx.RuleBuilder.CTMatching.match_ct_bytes/3.

match_ct_direction(builder, direction)

See NFTablesEx.RuleBuilder.CTMatching.match_ct_direction/2.

match_ct_helper(builder, helper)

See NFTablesEx.RuleBuilder.CTMatching.match_ct_helper/2.

match_ct_label(builder, label)

See NFTablesEx.RuleBuilder.CTMatching.match_ct_label/2.

match_ct_original_daddr(builder, addr)

See NFTablesEx.RuleBuilder.CTMatching.match_ct_original_daddr/2.

match_ct_original_saddr(builder, addr)

See NFTablesEx.RuleBuilder.CTMatching.match_ct_original_saddr/2.

match_ct_packets(builder, op, packets)

See NFTablesEx.RuleBuilder.CTMatching.match_ct_packets/3.

match_ct_state(builder, states)

See NFTablesEx.RuleBuilder.CTMatching.match_ct_state/2.

match_ct_status(builder, statuses)

See NFTablesEx.RuleBuilder.CTMatching.match_ct_status/2.

match_ct_zone(builder, zone)

See NFTablesEx.RuleBuilder.CTMatching.match_ct_zone/2.

match_dest_ip(builder, ip)

See NFTablesEx.RuleBuilder.IPMatching.match_dest_ip/2.

match_dest_mac(builder, mac)

See NFTablesEx.RuleBuilder.Layer2Matching.match_dest_mac/2.

match_dest_port(builder, port)

See NFTablesEx.RuleBuilder.PortMatching.match_dest_port/2.

match_dscp(builder, dscp)

See NFTablesEx.RuleBuilder.AdvancedMatching.match_dscp/2.

match_esp_spi(builder, spi)

See NFTablesEx.RuleBuilder.AdvancedMatching.match_esp_spi/2.

match_fragmented(builder, is_fragmented)

See NFTablesEx.RuleBuilder.AdvancedMatching.match_fragmented/2.

match_hoplimit(builder, op, hoplimit)

See NFTablesEx.RuleBuilder.TCPMatching.match_hoplimit/3.

match_icmp_code(builder, code)

See NFTablesEx.RuleBuilder.AdvancedMatching.match_icmp_code/2.

match_icmp_type(builder, type)

See NFTablesEx.RuleBuilder.AdvancedMatching.match_icmp_type/2.

match_icmpv6_code(builder, code)

See NFTablesEx.RuleBuilder.AdvancedMatching.match_icmpv6_code/2.

match_icmpv6_type(builder, type)

See NFTablesEx.RuleBuilder.AdvancedMatching.match_icmpv6_type/2.

match_iif(builder, ifname)

See NFTablesEx.RuleBuilder.Layer2Matching.match_iif/2.

match_length(builder, op, length)

See NFTablesEx.RuleBuilder.TCPMatching.match_length/3.

match_mark(builder, mark)

See NFTablesEx.RuleBuilder.AdvancedMatching.match_mark/2.

match_oif(builder, ifname)

See NFTablesEx.RuleBuilder.Layer2Matching.match_oif/2.

match_pkttype(builder, pkttype)

See NFTablesEx.RuleBuilder.AdvancedMatching.match_pkttype/2.

match_port_range(builder, min_port, max_port)

See NFTablesEx.RuleBuilder.PortMatching.match_port_range/3.

match_priority(builder, op, priority)

See NFTablesEx.RuleBuilder.AdvancedMatching.match_priority/3.

match_protocol(builder, protocol)

See NFTablesEx.RuleBuilder.TCPMatching.match_protocol/2.

match_set(builder, set_name, match_type)

See NFTablesEx.RuleBuilder.AdvancedMatching.match_set/3.

match_skgid(builder, gid)

See NFTablesEx.RuleBuilder.AdvancedMatching.match_skgid/2.

match_skuid(builder, uid)

See NFTablesEx.RuleBuilder.AdvancedMatching.match_skuid/2.

match_source_ip(builder, ip)

See NFTablesEx.RuleBuilder.IPMatching.match_source_ip/2.

match_source_mac(builder, mac)

See NFTablesEx.RuleBuilder.Layer2Matching.match_source_mac/2.

match_source_port(builder, port)

See NFTablesEx.RuleBuilder.PortMatching.match_source_port/2.

match_source_port_range(builder, min_port, max_port)

See NFTablesEx.RuleBuilder.PortMatching.match_source_port_range/3.

match_tcp_flags(builder, flags, mask)

See NFTablesEx.RuleBuilder.TCPMatching.match_tcp_flags/3.

match_ttl(builder, op, ttl)

See NFTablesEx.RuleBuilder.TCPMatching.match_ttl/3.

match_udp_dport(builder, port)

See NFTablesEx.RuleBuilder.PortMatching.match_udp_dport/2.

match_udp_port_range(builder, min_port, max_port)

See NFTablesEx.RuleBuilder.PortMatching.match_udp_port_range/3.

match_udp_source_port_range(builder, min_port, max_port)

See NFTablesEx.RuleBuilder.PortMatching.match_udp_source_port_range/3.

match_udp_sport(builder, port)

See NFTablesEx.RuleBuilder.PortMatching.match_udp_sport/2.

match_vlan_id(builder, vlan_id)

See NFTablesEx.RuleBuilder.Layer2Matching.match_vlan_id/2.

match_vlan_pcp(builder, pcp)

See NFTablesEx.RuleBuilder.Layer2Matching.match_vlan_pcp/2.

new(pid, table, chain, opts \\ [])

@spec new(pid(), String.t(), String.t(), keyword()) :: t()

Start building a new rule.

Parameters

pid - NFTex process

table - Table name

chain - Chain name

opts - Options:

:family - Protocol family (default: :inet)

Example

# Create builder with default IPv4/IPv6 family
builder = RuleBuilder.new(pid, "filter", "INPUT")

# Create builder for IPv6
builder = RuleBuilder.new(pid, "filter", "INPUT", family: :inet6)

notrack(builder)

See NFTablesEx.RuleBuilder.Verdicts.notrack/1.

queue_to_userspace(builder, queue_num)

See NFTablesEx.RuleBuilder.Verdicts.queue_to_userspace/2.

queue_to_userspace(builder, queue_num, opts)

See NFTablesEx.RuleBuilder.Verdicts.queue_to_userspace/3.

rate_limit(builder, rate, unit)

See NFTablesEx.RuleBuilder.Actions.rate_limit/3.

rate_limit(builder, rate, unit, opts)

See NFTablesEx.RuleBuilder.Actions.rate_limit/4.

redirect_to(builder, port)

See NFTablesEx.RuleBuilder.NAT.redirect_to/2.

reject(builder)

See NFTablesEx.RuleBuilder.Verdicts.reject/1.

reject(builder, type)

See NFTablesEx.RuleBuilder.Verdicts.reject/2.

restore_mark(builder)

See NFTablesEx.RuleBuilder.Actions.restore_mark/1.

return_from_chain(builder)

See NFTablesEx.RuleBuilder.Verdicts.return_from_chain/1.

save_mark(builder)

See NFTablesEx.RuleBuilder.Actions.save_mark/1.

set_connmark(builder, mark)

See NFTablesEx.RuleBuilder.Actions.set_connmark/2.

set_ct_helper(builder, helper)

See NFTablesEx.RuleBuilder.Actions.set_ct_helper/2.

set_ct_label(builder, label)

See NFTablesEx.RuleBuilder.Actions.set_ct_label/2.

set_ct_zone(builder, zone)

See NFTablesEx.RuleBuilder.Actions.set_ct_zone/2.

set_dscp(builder, dscp)

See NFTablesEx.RuleBuilder.Actions.set_dscp/2.

set_hoplimit(builder, hoplimit)

See NFTablesEx.RuleBuilder.Actions.set_hoplimit/2.

set_mark(builder, mark)

See NFTablesEx.RuleBuilder.Actions.set_mark/2.

set_tcp_mss(builder, mss)

See NFTablesEx.RuleBuilder.Verdicts.set_tcp_mss/2.

set_ttl(builder, ttl)

See NFTablesEx.RuleBuilder.Actions.set_ttl/2.

snat_to(builder, ip)

See NFTablesEx.RuleBuilder.NAT.snat_to/2.

snat_to(builder, ip, opts)

See NFTablesEx.RuleBuilder.NAT.snat_to/3.

synproxy(builder)

See NFTablesEx.RuleBuilder.Verdicts.synproxy/1.

synproxy(builder, opts)

See NFTablesEx.RuleBuilder.Verdicts.synproxy/2.

to_nft_command(builder)

@spec to_nft_command(t()) :: binary()

Build the nft command string without executing.

Converts the JSON expression list to human-readable nft syntax. Returns the nft command string that would be sent to add this rule. Useful for batching, remote execution, or inspection.

Returns

nft command string

Examples

# Build command
cmd = RuleBuilder.new(pid, "filter", "INPUT")
|> RuleBuilder.match_source_ip("192.168.1.100")
|> RuleBuilder.drop()
|> RuleBuilder.to_nft_command()
#=> "add rule inet filter INPUT ip saddr 192.168.1.100 drop"

# Use in batch
cmd1 = RuleBuilder.new(pid, "filter", "INPUT")
       |> match_source_ip("1.2.3.4")
       |> drop()
       |> to_nft_command()

cmd2 = RuleBuilder.new(pid, "filter", "INPUT")
       |> match_source_ip("5.6.7.8")
       |> drop()
       |> to_nft_command()

batch = Batch.new()
|> Batch.add(cmd1)
|> Batch.add(cmd2)
|> Batch.execute()

# Send to remote node
cmd = RuleBuilder.new(pid, "filter", "INPUT")
      |> match_dest_port(22)
      |> rate_limit(10, :minute)
      |> accept()
      |> to_nft_command()

MyTransport.send_to_node("firewall-1", cmd)